+++ src/lib/smartpls.h
@@ -12,9 +12,11 @@
 #include <stdbool.h>
 
 bool smartpls_save_sticker(sds workdir, const char *playlist, const char *sticker,
-    int max_entries, int min_value, const char *sort, bool sortdesc);
-bool smartpls_save_newest(sds workdir, const char *playlist, int timerange, const char *sort, bool sortdesc);
-bool smartpls_save_search(sds workdir, const char *playlist, const char *expression, const char *sort, bool sortdesc);
+        const char *value, const char *op, const char *sort, bool sortdesc, int max_entries);
+bool smartpls_save_newest(sds workdir, const char *playlist, int timerange,
+        const char *sort, bool sortdesc, int max_entries);
+bool smartpls_save_search(sds workdir, const char *playlist, const char *expression,
+        const char *sort, bool sortdesc, int max_entries);
 
 bool smartpls_update(const char *playlist);
 bool smartpls_update_all(void);