+++ src/mympd_api/extra_media.h
@@ -4,10 +4,11 @@
  https://github.com/jcorporation/mympd
 */
 
-#ifndef MYMPD_API_UTILITY_H
-#define MYMPD_API_UTILITY_H
+#ifndef MYMPD_API_EXTRA_MEDIA_H
+#define MYMPD_API_EXTRA_MEDIA_H
 
 #include "src/lib/mympd_state.h"
 
 sds mympd_api_get_extra_media(struct t_mpd_state *mpd_state, sds buffer, const char *uri, bool is_dirname);
+
 #endif